+++ kie-wb-common-screens/kie-wb-common-project-editor/kie-wb-common-project-editor-backend/src/main/java/org/kie/workbench/common/screens/projecteditor/backend/server/ProjectScreenModelLoader.java
@@ -33,7 +33,8 @@ import org.uberfire.backend.server.util.Paths;
 import org.uberfire.backend.vfs.Path;
 
 import static java.util.stream.Collectors.toList;
-import static org.kie.workbench.common.screens.projecteditor.model.ProjectScreenModel.GitUrl;
+
+import org.kie.workbench.common.screens.projecteditor.model.GitUrl;
 
 public class ProjectScreenModelLoader {
